Disable deprecation type tests in the browser
We don't have a good way to make the YAML file accessible to tests
running in a browser, so we can just skip them there.
